Introduced inheritance for class "OggettoCeleste"
The main reason for this commit is the inheritance; anyway, it also
deals with a better-engineered approach of the communication between
classes: DAO and controllers are no more static classes, so it's now
possible to explore dependencies between these classes. Another issue
solved relates to methods which had as input/output ArrayList<T> : I
changed it to List<T> to provide the possibility to change type of list
without changing all components calling these methods.